pdf-to-text-urdu/requirements.txt
absl-py==1.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
astor==0.8.1
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs==22.2.0
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
cached-property==1.5.2
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814693874/work
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
dataclasses==0.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
Django==3.2.25
djangorestframework==3.15.1
doxapy==0.9.2
drf-yasg==1.21.8
entrypoints==0.3
et-xmlfile==1.1.0
gast==0.6.0
grpcio==1.48.2
h5py==3.1.0
idna==3.10
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.5.1
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///tmp/build/80754af9/jedi_1606932572482/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
joblib==1.1.1
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core==4.9.2
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528150516/work
matplotlib==3.3.4
mistune==0.8.4
nbclient==0.5.9
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914804165/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio==1.6.0
nltk==3.6.7
notebook==6.4.10
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1626681920064/work
openpyxl==3.1.3
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.5
pandocfilters==1.5.1
parso==0.7.0
path.py==8.1.2
pdf2image==1.17.0
pexpect==4.9.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.14.0
pyparsing==3.1.2
pyrsistent==0.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.8.8
requests==2.27.1
Send2Trash==1.8.3
simplegeneric==0.8.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
sqlparse==0.4.4
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.6.0
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
tqdm==4.64.1
traitlets @ file:///tmp/build/80754af9/traitlets_1632746497744/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
uritemplate==4.1.1
urllib3==1.26.20
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==2.0.3
xlrd==2.0.1
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
